Format changes to the info cards and signup button
Added some formatting to the info cards makes it stretch to a
certain width and grow/shrink margins between info cards as the page
shrinks and grows. However, using align-items: stretch makes the cards
align to the left. A fixed margin may be best in order to keep the cards
in one spot.